+++ crates/test-ui/src/main.rs
@@ -13,8 +13,14 @@
 
 fn main() {
     let t = trybuild::TestCases::new();
+
     let path = std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
         .join("ui")
         .join("*.rs");
     t.compile_fail(path);
+
+    let path = std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
+        .join("ui-ignore")
+        .join("*.rs");
+    t.pass(path);
 }